PMLDL_ASS1/Code/Deployment/api/requirements.txt
fastapi
uvicorn
pydantic
torch
torchvision
Pillow
tqdm
matplotlib
pybase64